+++ src/main/java/kitchenpos/menu/domain/MenuRepository.java
@@ -1,4 +1,6 @@
-package kitchenpos.domain;
+package kitchenpos.menu.domain;
+
+import kitchenpos.menu.domain.Menu;
 
 import java.util.List;
 import java.util.Optional;
