+++ src/test/java/kitchenpos/Fixtures.java
@@ -1,6 +1,6 @@
 package kitchenpos;
 
-import kitchenpos.domain.Menu;
+import kitchenpos.menu.domain.Menu;
 import kitchenpos.domain.MenuGroup;
 import kitchenpos.domain.MenuProduct;
 import kitchenpos.domain.Order;
